refactor: main.js의 컴포넌트를 개별 파일로 분리
