+++ src/validateRestaurant.test.js
@@ -1,5 +1,5 @@
 import { ERROR_MESSAGE } from "./constants/constants.js";
-import { validateRestaurant } from "./validateRestaurant.js";
+import validateRestaurant from "./validateRestaurant.js";
 
 describe("음식점 이름 테스트", () => {
   const restaurantsNameList = [
